pyproject.toml
[tool.poetry]
name = "django-celery"
version = "0.1.0"
description = ""
authors = ["Yaroslav Pekatoros <y.pekatoros@gmail.com>"]
readme = "README.md"
packages = [{include = "django_celery"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
celery = "^5.2.7"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
